PVP-256 Update README; Move route components to different file; Rename route lists
